+++ src/main/java/com/example/dao/UsuarioDao.java
@@ -5,9 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UsuarioDao extends JpaRepository <Usuario, Integer> {
+public interface UsuarioDao extends JpaRepository <Usuario, Long> {
     Usuario findBynombreUsuario(String username);
     List<Usuario> findByRol_NombreRol(String nomRol);
+
 
 
 }
