+++ src/main/java/com/example/servicio/MaterialServicio.java
@@ -1,13 +1,15 @@
 package com.example.servicio;
 
 import com.example.domain.Material;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 
 public interface MaterialServicio {
 
     List<Material> listarTodos();
-    Material obtenerPorId(Integer idMaterial);
+    Material obtenerPorId(Long idMaterial);
     void guardar(Material material);
-    void eliminar(Integer idMaterial);
+    void eliminar(Long idMaterial);
 }
